requirements.txt
asgiref==3.3.0
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.0.11
defusedxml==0.6.0
dj-rest-auth==1.1.2
Django==3.1.3
django-allauth==0.43.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
drf-generators==0.5.0
google-api-core==1.25.1
google-api-python-client==1.12.8
google-apis-oauth-django==0.1.0
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
imageio==2.1.2
jsonpickle==0.9.6
mysqlclient==2.0.1
oauth2client==4.1.2
oauthlib==3.1.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
tqdm==4.11.2
uritemplate==3.0.1
urllib3==1.25.11
